+++ app/controllers/admin/events_controller.rb
@@ -13,6 +13,7 @@ class Admin::EventsController < Admin::BaseController
     @event = Event.new(event_params)
 
     if @event.save
+      handle_image_uploads(@event)
       redirect_to admin_events_path, notice: "Event created"
     else
       flash.now[:alert] = @event.errors.full_messages.to_sentence.presence || "Event creation failed"
@@ -25,6 +26,7 @@ class Admin::EventsController < Admin::BaseController
 
   def update
     if @event.update(event_params)
+      handle_image_uploads(@event)
       redirect_to admin_events_path, notice: "Event updated"
     else
       flash.now[:alert] = @event.errors.full_messages.to_sentence.presence || "Event update failed"
@@ -44,6 +46,31 @@ class Admin::EventsController < Admin::BaseController
   end
 
   def event_params
-    params.fetch(:event, {}).permit(:title, :event_date, :location, :description, :link)
+    params.fetch(:event, {}).permit(:title, :event_date, :location, :description, :link, :image_key, :image_alt_key)
+  end
+
+  def handle_image_uploads(event)
+    upload_event_image(event, params.dig(:event, :image), "main")
+    upload_event_image(event, params.dig(:event, :alt_image), "alt")
+  end
+
+  def upload_event_image(event, uploaded, image_type)
+    return if uploaded.blank?
+
+    ext = File.extname(uploaded.original_filename)
+    key = "events/#{event.id}/#{image_type}#{ext}"
+
+    begin
+      S3Service.new.upload(uploaded, key)
+    rescue StandardError => e
+      Rails.logger.error("S3 upload failed for event #{event.id}: #{e.message}")
+      return
+    end
+
+    if image_type == "main"
+      event.update!(image_key: key)
+    else
+      event.update!(image_alt_key: key)
+    end
   end
 end